refactor: AI 패널 API 레이어를 feature 내부로 이동
