+++ apps/client/src/features/ai-panel/apis/query-key.ts
