+++ apps/client/src/features/ai-panel/apis/type.ts
