+++ apps/client/src/features/ai-panel/hooks/use-ai-panel-chat.test.tsx
@@ -9,7 +9,7 @@ const queryMocks = vi.hoisted(() => ({
   useSaveAiMemo: vi.fn(),
 }));
 
-vi.mock('@shared/apis/prompt/queries', () => queryMocks);
+vi.mock('../apis/queries', () => queryMocks);
 
 import { useAiPanelChat } from './use-ai-panel-chat';
